requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.0
Django==3.2
django-celery-beat==2.5.0
django-cors-headers==4.2.0
django-environ==0.10.0
django-filter==23.2
django-grappelli==3.0.7
django-jazzmin==2.6.0
django-jet==1.0.8
django-jet-reboot==1.3.3
django-phonenumber-field==7.1.0
django-timezone-field==5.1
djangorestframework==3.13.0
drf-yasg2==1.19.4
frozenlist==1.4.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.1
magic-filter==1.0.11
MarkupSafe==2.1.3
multidict==6.0.4
packaging==23.1
phonenumbers==8.13.19
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==5.0.0
requests==2.31.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.0
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
